Initial sketch.
This provides a _very_ minimal (and not yet functional) bootloader
containing a plausible initial set of functionality. It calls the ROM
API to validate each of two image slots, and chooses a valid image or
panics.

It does a reasonably complete reset of machine state before jumping into
the new image, though some TODOs persist.

On Rust 1.65 the release build is 332 bytes of code; plus the vector
table this gives a total flash size of 636 bytes.

Currently the program does not use (non-stack) RAM.