Write a CRC32 implementation for size reasons.
This knocks a bit over 400 bytes, or 26%, off the size of the program
in exchange for decreased CRC performance (since it no longer relies on
a precomputed lookup table).

In this application it's hard to say _how_ decreased the performance is:
ARM's combined-XOR-and-shift instructions make CRC32 pretty concise.
Given that we're targeting a 96MHz part and only need to CRC, at most, a
couple of hundred kiB once per boot, I'm sure performance will be fine.