Add a HardFault handler.
This makes sure we indicate fault on an external pin, whether it's a
software fault (panic) or a hardware fault (e.g. a bus error).

By some careful rearranging, I was able to do this without adding to the
size of the compiled program.

We are unfortunately still paying for the cortex-m-rt
HardFaultTrampoline routine. See:

    https://github.com/rust-embedded/cortex-m/issues/406