Fix slot-size math error.
The validation logic was rejecting any image larger than 1/4 the slot
size, because it was comparing the image's self-declared length in bytes
to our measure of the length in _words._

The slot size was still the value for the original partition layout,
which I've fixed; this was less of an issue since using it would require
having written a properly signed image across two firmware slots. But
still, good to fix.