Record which transient override was observed.
This changes the "acknowledge transient override" logic to set the first
byte to 0 if the override indicated slot A, and 1 if it indicated slot
B.

This should allow a program to detect that a transient override was
processed (because it's now zeros) and what override was found (based on
the low order bit). A program can further compare this to its physical
slot location to determine whether the override was able to be honored.

Fixes #28.