rot_carrier: fix floating override buttons
If you have the Digilent "four pushbuttons" PMOD plugged into the Aux
port, it gives you the ability to override boot selection.

However, if you _don't_ have it plugged in, the pins are floating, and
the boot gets overridden more-or-less at random. That's obviously not
ideal, so this change turns on pulldown resistors to ensure that the
floating pins are in a predictable (and not-overridey) state.