Fix off by one when checking for programmed pages (#46)
We should be adding `BYTES_PER_FWORD_U32 - 1`. This means the
calculation is incorrect for lengths that are divisible by 0 or 15.
and we'll be checking one extra word. This typically goes unnoticed
since we programm full pages. If the image length is exactly a
page size we will incorrectly expect one extra programmed page.